requirements.txt
fastapi
uvicorn
celery
redis
pydantic
PyPDF2
python-multipart
python-dotenv
requests
reportlab
